♻️ [Refactor] events 모듈 도입 및 listener 초기화
